# api.yml
...
openapi: 3.0.1
info:
  title: GeoSupport_ms
  description: |-
    Hello world, and welcome to the GeoSupport_ms API gide!
    GeoSupport is an ASP.Net Core microservice for the GeoSmart project.
    Go ahead and check all the request you can use :) .
  version: v1
paths:
  "/api/Color":
    get:
      tags:
      - Color
      summary: Gets a list of Color elements
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color"
    post:
      tags:
      - Color
      summary: Inserts a new Color register in database
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Color"
          text/json:
            schema:
              "$ref": "#/components/schemas/Color"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Color"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color"
  "/api/Color/{id}":
    get:
      tags:
      - Color
      summary: Gets a specific Color element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color"
        '404':
          description: If there is not item whit that id
    put:
      tags:
      - Color
      summary: Updates a specific Color element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Color"
          text/json:
            schema:
              "$ref": "#/components/schemas/Color"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Color"
      responses:
        '200':
          description: Success
        '400':
          description: If given id not valid
        '404':
          description: When an error is producted updating db
    delete:
      tags:
      - Color
      summary: Deletes a specific Color register given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color"
  "/api/Color_Flag":
    get:
      tags:
      - Color_Flag
      summary: Gets a list of Color_Flags elements
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color_Flag"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color_Flag"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Color_Flag"
    post:
      tags:
      - Color_Flag
      summary: Inserts a new Color_Flag register in database
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
          text/json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
        '409':
          description: When there is already a register with those Ids
  "/api/Color_Flag/{c_id}/{f_id}":
    get:
      tags:
      - Color_Flag
      summary: Gets a specific Color_Flag element given its color_id (c_id) and flag_id
        (f_id)
      parameters:
      - name: c_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: f_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
        '404':
          description: If there is not item whit those ids
    put:
      tags:
      - Color_Flag
      summary: Updates a specific Color_Flag element given its color_id (c_id) and
        flag_id (f_id)
      parameters:
      - name: c_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: f_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
          text/json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Color_Flag"
      responses:
        '200':
          description: Success
        '400':
          description: If given ids not valid
        '404':
          description: When an error is producted updating db
    delete:
      tags:
      - Color_Flag
      summary: Deletes a specific Color_Flag element given its color_id (c_id) and
        flag_id (f_id)
      parameters:
      - name: c_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: f_id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Color_Flag"
        '404':
          description: If there is not item whit those ids
...
  "/api/Flag/{id}":
    get:
      tags:
      - Flag
      summary: Gets a specific Flag element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/FlagColors"
            application/json:
              schema:
                "$ref": "#/components/schemas/FlagColors"
            text/json:
              schema:
                "$ref": "#/components/schemas/FlagColors"
        '404':
          description: If there is not item whit that id
    put:
      tags:
      - Flag
      summary: Updates a specific Flag element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          text/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Flag"
      responses:
        '200':
          description: Success
        '400':
          description: If given id not valid
        '404':
          description: When an error is producted updating db
    delete:
      tags:
      - Flag
      summary: Deletes a specific Flag register given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Flag"
        '404':
          description: If there is not item whit that id
...
components:
  schemas:
    Flag:
      required:
      - flagImage
      type: object
      properties:
        id_flag:
          type: integer
          format: int32
        flagImage:
          maxLength: 2000
          type: string
        colors:
          type: array
          items:
            "$ref": "#/components/schemas/Color"
          nullable: true
      additionalProperties: false
    Color:
      required:
      - hexCode
      - name
      type: object
      properties:
        id_color:
          type: integer
          format: int32
        name:
          maxLength: 20
          type: string
        hexCode:
          maxLength: 20
          type: string
        flags:
          type: array
          items:
            "$ref": "#/components/schemas/Flag"
          nullable: true
      additionalProperties: false
    Color_Flag:
      required:
      - order
      type: object
      properties:
        id_flag:
          type: integer
          format: int32
        id_color:
          type: integer
          format: int32
        order:
          type: integer
          format: int32
        color:
          "$ref": "#/components/schemas/Color"
        flag:
          "$ref": "#/components/schemas/Flag"
      additionalProperties: false
...
    ColorOrder:
      type: object
      properties:
        id_color:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        hexCode:
          type: string
          nullable: true
        order:
          type: integer
          format: int32
      additionalProperties: false
    FlagColors:
      type: object
      properties:
        id_flag:
          type: integer
          format: int32
        flagImage:
          type: string
          nullable: true
        colors:
          type: array
          items:
            "$ref": "#/components/schemas/ColorOrder"
          nullable: true
      additionalProperties: false